Eliminacion modelos FAQ y ACtualización, TRyCAtch creado en cascada, eliminación de la capa repository.
