--- CrowdSisters/Models/Actualizacion.cs
@@ -1,31 +1,0 @@
-﻿using System;
-using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
-
-namespace CrowdSisters.Models
-{
-    [Table("Actualizacion")]
-    public class Actualizacion
-    {
-        [Key]
-        [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int IDActualizacion { get; set; }
-
-        [Required]
-        [ForeignKey("Proyecto")]
-        public int FKProyecto { get; set; }
-
-        public Proyecto Proyecto { get; set; }
-
-        [Required]
-        [StringLength(50)]
-        public string Titulo { get; set; }
-
-        [Required]
-        [StringLength(200)]
-        public string Contenido { get; set; }
-
-        [Required]
-        public DateTime FechaActualizacion { get; set; }
-    }
-}
