+++ CrowdSisters/Program.cs
@@ -1,6 +1,5 @@
 using CrowdSisters.Conections;
 using CrowdSisters.DAL;
-using CrowdSisters.Repository;
 using CrowdSisters.Services;
 
 var builder = WebApplication.CreateBuilder(args);
@@ -11,7 +10,6 @@ builder.Services.AddControllersWithViews();
 // Register the Connection class with dependency injection
 builder.Services.AddScoped<Connection>();
 builder.Services.AddScoped<DALProyecto>();
-builder.Services.AddScoped<RepoProyecto>();
 builder.Services.AddScoped<ServiceProyecto>();
 
 var app = builder.Build();
