--- CrowdSisters/Repository/RepoProyecto.cs
@@ -1,88 +1,0 @@
-﻿using System;
-using System.Threading.Tasks;
-using CrowdSisters.DAL;
-using CrowdSisters.Models;
-
-namespace CrowdSisters.Repository
-{
-    public class RepoProyecto
-    {
-        private readonly DALProyecto _dalProyecto;
-
-        public RepoProyecto(DALProyecto dalProyecto)
-        {
-            _dalProyecto = dalProyecto;
-        }
-
-        public async Task<bool> CreateProyectoAsync(Proyecto proyecto)
-        {
-            try
-            {
-                return await _dalProyecto.CreateAsync(proyecto);
-            }
-            catch (Exception ex)
-            {
-                // Manejo de errores (log, rethrow, etc.)
-                Console.WriteLine($"Error al crear proyecto: {ex.Message}");
-                return false;
-            }
-        }
-
-        public async Task<Proyecto> GetProyectoByIdAsync(int id)
-        {
-            try
-            {
-                return await _dalProyecto.GetByIdAsync(id);
-            }
-            catch (Exception ex)
-            {
-                // Manejo de errores
-                Console.WriteLine($"Error al obtener proyecto: {ex.Message}");
-                return null;
-            }
-        }
-
-        public async Task<bool> UpdateProyectoAsync(Proyecto proyecto)
-        {
-            try
-            {
-                return await _dalProyecto.UpdateAsync(proyecto);
-            }
-            catch (Exception ex)
-            {
-                // Manejo de errores
-                Console.WriteLine($"Error al actualizar proyecto: {ex.Message}");
-                return false;
-            }
-        }
-
-        public async Task<bool> DeleteProyectoAsync(int id)
-        {
-            try
-            {
-                return await _dalProyecto.DeleteAsync(id);
-            }
-            catch (Exception ex)
-            {
-                // Manejo de errores
-                Console.WriteLine($"Error al eliminar proyecto: {ex.Message}");
-                return false;
-            }
-        }
-
-        // Obtener todos los proyectos
-        public async Task<IEnumerable<Proyecto>> GetAllProyectosAsync()
-        {
-            try
-            {
-                return await _dalProyecto.GetAllAsync();
-            }
-            catch (Exception ex)
-            {
-                // Manejo de errores
-                Console.WriteLine($"Error al obtener todos los proyectos: {ex.Message}");
-                return new List<Proyecto>(); // Retorna una lista vacía en caso de error
-            }
-        }
-    }
-}
