+++ CrowdSisters/Views/Proyecto/Error.cshtml
@@ -1,0 +1,6 @@
+﻿@{
+    ViewData["Title"] = "Error";
+}
+
+<h1>Error</h1>
+<p>@ViewBag.ErrorMessage</p>
